writing test cases
