+++ component/src/test/java/io/siddhi/extension/store/mongodb/JoinMongoTableTest.java
@@ -317,254 +317,4 @@ public class JoinMongoTableTest {
 
         Assert.assertEquals(eventCount.intValue(), 1, "Read events failed");
     }
-
-    @Test
-    public void testMongoTableJoinQuery7() throws InterruptedException {
-        log.info("testMongoTableJoinQuery7");
-
-        MongoTableTestUtils.dropCollection(uri, "FooTable");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String streams = "" +
-                "define stream StockStream (symbol string, price float); " +
-                "define stream FooStream (symbol string, volume int, name string); " +
-                "@store(type = 'mongodb' , mongodb.uri='" + uri + "')" +
-                "define table FooTable (symbol string, price float);";
-        String query = "" +
-                "@info(name = 'query1') " +
-                "from StockStream " +
-                "insert into FooTable ;" +
-                "" +
-                "@info(name = 'query2') " +
-                "from FooStream as s join FooTable as t " +
-                "on s.symbol != t.symbol "+
-                "select s.symbol as streamSymbol, t.symbol as tblSymbol, t.price + t.price as price, '100' as fixedVal, s.volume + s.volume as streamVolume, s.name as streamName " +
-                "having price < 12 and streamVolume > 1 or streamName == 'siddhi' " +
-                "order by streamVolume "+
-                "limit 1 "+
-                "offset 1 "+
-                "insert into OutputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("query2", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                if (inEvents != null) {
-                    for (Event event : inEvents) {
-                        eventCount.incrementAndGet();
-                        switch (eventCount.intValue()) {
-                            case 1:
-                                Assert.assertEquals(new Object[]{"WSO2","IBM",8.5,100,10,"siddhi"}, event.getData());
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-                }
-            }
-
-        });
-
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
-        InputHandler fooStream = siddhiAppRuntime.getInputHandler("FooStream");
-        siddhiAppRuntime.start();
-
-        stockStream.send(new Object[]{"WSO2", 6.5f});
-        stockStream.send(new Object[]{"WSO2", 6.5f});
-        stockStream.send(new Object[]{"IBM", 9.5f});
-        stockStream.send(new Object[]{"IBM", 8.5f});
-        fooStream.send(new Object[]{"WSO2",10,"siddhi"});
-        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
-
-        siddhiAppRuntime.shutdown();
-
-        Assert.assertEquals(eventCount.intValue(), 1, "Read events failed");
-    }
-
-    @Test
-    public void testMongoTableJoinQuery8() throws InterruptedException {
-        log.info("testMongoTableJoinQuery8");
-
-        MongoTableTestUtils.dropCollection(uri, "FooTable");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String streams = "" +
-                "define stream StockStream (symbol string, price float, weight int); " +
-                "define stream FooStream (symbol string, volume int); " +
-                "@store(type = 'mongodb' , mongodb.uri='" + uri + "')" +
-                "define table FooTable (symbol string, price float, weight int);";
-        String query = "" +
-                "@info(name = 'query1') " +
-                "from StockStream " +
-                "insert into FooTable ;" +
-                "" +
-                "@info(name = 'query2') " +
-                "from FooStream as s join FooTable as t " +
-                "on t.symbol == s.symbol "+
-                "select t.symbol, avg(t.weight) as avgweight, min(t.price) as minprice, max(t.price) as maxprice, sum(t.price) as sumprice " +
-                "group by t.symbol " +
-                "having maxprice > 9 "+
-                "order by maxprice "+
-                "limit 5 "+
-                "insert into OutputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("query2", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                if (inEvents != null) {
-                    for (Event event : inEvents) {
-                        eventCount.incrementAndGet();
-                        switch (eventCount.intValue()) {
-                            case 1:
-                                Assert.assertEquals(new Object[]{"GOOGLE", 13.0, 9.5, 12.5, 22.0}, event.getData());
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-                }
-            }
-
-        });
-
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
-        InputHandler fooStream = siddhiAppRuntime.getInputHandler("FooStream");
-        siddhiAppRuntime.start();
-
-        stockStream.send(new Object[]{"GOOGLE", 12.5f, 10});
-        stockStream.send(new Object[]{"APPLE", 10.5f, 14});
-        stockStream.send(new Object[]{"GOOGLE", 9.5f, 16});
-        stockStream.send(new Object[]{"APPLE", 8.5f, 18});
-        fooStream.send(new Object[]{"GOOGLE",10});
-        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
-
-        siddhiAppRuntime.shutdown();
-
-        Assert.assertEquals(eventCount.intValue(), 1, "Read events failed");
-    }
-
-    @Test
-    public void testMongoTableJoinQuery9() throws InterruptedException {
-        log.info("testMongoTableJoinQuery9");
-        //Object reads
-
-        MongoTableTestUtils.dropCollection(uri, "FooTable");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String streams = "" +
-                "define stream StockStream (symbol string, price float); " +
-                "define stream FooStream (symbol string, volume int); " +
-                "@store(type = 'mongodb' , mongodb.uri='" + uri + "')" +
-                "define table FooTable (symbol string, price float);";
-        String query = "" +
-                "@info(name = 'query1') " +
-                "from StockStream " +
-                "insert into FooTable ;" +
-                "" +
-                "@info(name = 'query2') " +
-                "from FooStream as s join FooTable as t " +
-                "select t.symbol as tblSymbol, t.price - s.volume + 6 / 2 as addVars " +
-                "order by tblSymbol "+
-                "limit 2 "+
-                "offset 1 "+
-                "insert into OutputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("query2", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                if (inEvents != null) {
-                    for (Event event : inEvents) {
-                        eventCount.incrementAndGet();
-                        switch (eventCount.intValue()) {
-                            case 1:
-                                Assert.assertEquals(new Object[]{"WSO2","IBM",8.5,100,"10","prabod"}, event.getData());
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-                }
-            }
-
-        });
-
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
-        InputHandler fooStream = siddhiAppRuntime.getInputHandler("FooStream");
-        siddhiAppRuntime.start();
-
-        stockStream.send(new Object[]{"WSO2", 20f});
-        stockStream.send(new Object[]{"GOOGLE", 20f});
-        stockStream.send(new Object[]{"WSO2", 30f});
-        stockStream.send(new Object[]{"IBM", 40f});
-        fooStream.send(new Object[]{"WSO2",10});
-        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
-
-        siddhiAppRuntime.shutdown();
-
-        Assert.assertEquals(eventCount.intValue(), 1, "Read events failed");
-    }
-
-    @Test
-    public void testMongoTableJoinQuery10() throws InterruptedException {
-        log.info("testMongoTableJoinQuery10");
-
-        MongoTableTestUtils.dropCollection(uri, "FooTable");
-
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String streams = "" +
-                "define stream StockStream (symbol string, price int, isFraud bool, isNewStock bool); " +
-                "define stream FooStream (symbol string, isChecked bool, isOlderStock bool); " +
-                "@store(type = 'mongodb' , mongodb.uri='" + uri + "')" +
-                "define table FooTable (symbol string, price int, isFraud bool, isNewStock bool);";
-        String query = "" +
-                "@info(name = 'query1') " +
-                "from StockStream " +
-                "insert into FooTable ;" +
-                "" +
-                "@info(name = 'query2') " +
-                "from FooStream as s join FooTable as t " +
-                "on s.symbol != t.symbol "+
-                "select t.symbol as tblSymbol, (t.isFraud or s.isChecked and s.isOlderStock and true) as isCheckedData, t.price as tblPrice " +
-                "having tblPrice > 5 and tblSymbol != 'WSO2' "+
-                "order by tblPrice DESC "+
-                "limit 2 "+
-                "offset 1 "+
-                "insert into OutputStream ;";
-
-        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
-        siddhiAppRuntime.addCallback("query2", new QueryCallback() {
-            @Override
-            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
-                if (inEvents != null) {
-                    for (Event event : inEvents) {
-                        eventCount.incrementAndGet();
-                        switch (eventCount.intValue()) {
-                            case 1:
-                                Assert.assertEquals(new Object[]{"LINUX", true, 10}, event.getData());
-                                break;
-                            default:
-                                break;
-                        }
-                    }
-                }
-            }
-
-        });
-
-        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
-        InputHandler fooStream = siddhiAppRuntime.getInputHandler("FooStream");
-        siddhiAppRuntime.start();
-
-        stockStream.send(new Object[]{"LINUX", 10, true, false});
-        stockStream.send(new Object[]{"GOOGLE", 12, false, true});
-        fooStream.send(new Object[]{"WSO2", true, false});
-        SiddhiTestHelper.waitForEvents(waitTime, 1, eventCount, timeout);
-
-        siddhiAppRuntime.shutdown();
-
-        Assert.assertEquals(eventCount.intValue(), 1, "Read events failed");
-    }
-
 }
